+++ pyout/field.py
@@ -319,6 +319,39 @@ class StyleProcessors(object):
             return self.render(style_attr or lookup_value, result)
         return proc
 
+    def by_re_lookup(self, style_key, style_value, re_flags=0):
+        """Return a processor for a "re_lookup" style value.
+
+        Parameters
+        ----------
+        style_key : str
+            A style key.
+        style_value : dict
+            A dictionary with a "re_lookup" style value that consists of a
+            sequence of items where each item should have the form `(regexp,
+            x)`, where regexp is a regular expression to match against the
+            field value and x is either a style attribute (str) and a boolean
+            flag indicating to use the style attribute named by `style_key`.
+        re_flags : int
+            Passed through as flags argument to re.compile.
+
+        Returns
+        -------
+        A function.
+        """
+        style_attr = style_key if self.style_types[style_key] is bool else None
+        regexps = [(re.compile(r, flags=re_flags), v)
+                   for r, v in style_value["re_lookup"]]
+
+        def proc(value, result):
+            for r, lookup_value in regexps:
+                if r.search(value):
+                    if not lookup_value:
+                        return result
+                    return self.render(style_attr or lookup_value, result)
+            return result
+        return proc
+
     def by_interval_lookup(self, style_key, style_value):
         """Return a processor for an "interval" style value.
 
@@ -394,6 +427,7 @@ class StyleProcessors(object):
 
         fns = {"simple": self.by_key,
                "lookup": self.by_lookup,
+               "re_lookup": self.by_re_lookup,
                "interval": self.by_interval_lookup}
 
         for key in self.style_types:
@@ -403,6 +437,9 @@ class StyleProcessors(object):
             vtype = value_type(column_style[key])
             fn = fns[vtype]
             args = [key, column_style[key]]
+            if vtype == "re_lookup":
+                args.append(sum(getattr(re, f)
+                                for f in column_style.get("re_flags", [])))
             yield fn(*args)
 
         yield flanks.join_flanks